+++ R/interpret.R
@@ -171,6 +171,8 @@ interpret.effectsize_table <- function(x, rules, ...) {
     # xtab
     Cramers_v = ,
     Cramers_v_adjusted = ,
+    normalized_chi = ,
+    Cohens_w = ,
     phi = ,
     phi_adjusted = interpret_cramers_v(value, rules = rules),
     Cohens_g = interpret_cohens_g(value, rules = rules),